+++ RELEASE-NOTES.md
@@ -21,7 +21,7 @@
 ## New additions
 
 ## Fixes and improvements
-
+* Fix certificate connection issues.
 
 # v3.7.0