+++ RELEASE-NOTES.md
@@ -21,6 +21,15 @@
 ## New additions
 
 ## Fixes and improvements
+
+
+# v3.7.1
+
+## Deprecations
+
+## New additions
+
+## Fixes and improvements
 * Fix certificate connection issues.
 * Fix `snow spcs image-registry login` slow query problem.